Refactor Wasm loader
